230319 Week4-STEP02 피드백 후 수정완료함. 100 근접값은 구현 못했음...ㅠ_ㅠ
